Added PostgreSQL and also Api Endpoints for users

+++ routers/todos.py
@@ -5,9 +5,13 @@ from models import Todos
 from database import SessionLocal
 from sqlalchemy.orm import Session
 from typing import Annotated
+from .auth import get_current_user
 
 
-router = APIRouter()
+router = APIRouter(
+    prefix='/todos',
+    tags=['TODOS']
+)
 
 def get_db():
     db = SessionLocal()
@@ -17,6 +21,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title : str = Field(min_length=3)
@@ -26,26 +31,33 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/",status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return  db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Unauthorized Access')
+    return  db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
-@router.get("/todo/{todo_id}",status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router.get("/read_todo/{todo_id}",status_code=status.HTTP_200_OK)
+async def read_todo(user:user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Unauthorized Access')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail='Todo Not Found')
 
-@router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
-
+@router.post("/create_todo", status_code=status.HTTP_201_CREATED)
+async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Unauthorized Access')
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
 
-@router.put("update_todo/{todo_id}/", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_id: int, todo_request: TodoRequest):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router.put("/update_todo/{todo_id}/", status_code=status.HTTP_204_NO_CONTENT)
+async def update_todo(user:user_dependency, db: db_dependency, todo_id: int, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Unauthorized Access')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo Not Found')
     todo_model.title = todo_request.title
@@ -57,8 +69,10 @@ async def update_todo(db: db_dependency, todo_id: int, todo_request: TodoRequest
     db.commit()
 
 @router.delete("/delete_todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user:user_dependency,db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Unauthorized Access')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         db.delete(todo_model)
         db.commit()
